Revert change log migration changes and remove affected files  / references
