+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsinterventionsservice/service/ReferralServiceTest.kt
@@ -97,7 +97,6 @@ class ReferralServiceTest @Autowired constructor(
   private val supplierAssessmentFactory = SupplierAssessmentFactory(entityManager)
   private val serviceUserDataFactory = ServiceUserFactory(entityManager)
   private val referralDetailsFactory = ReferralDetailsFactory(entityManager)
-  private val changeLogMigrationService = ChangeLogMigrationService(changelogRepository)
 
   private val referralEventPublisher: ReferralEventPublisher = mock()
   private val referralConcluder: ReferralConcluder = mock()
@@ -132,8 +131,7 @@ class ReferralServiceTest @Autowired constructor(
     hmppsAuthService,
     telemetryService,
     referralDetailsRepository,
-    changelogRepository,
-    changeLogMigrationService
+    changelogRepository
   )
 
   companion object {
